+++ worldwind/src/commonMain/kotlin/earth/worldwind/geom/TileMatrix.kt
@@ -30,13 +30,13 @@ open class TileMatrix internal constructor(
      * @param sector the desired sector to check tile count
      * @return Number of tiles which fit specified sector at this level
      */
-    fun tilesInSector(sector: Sector): Int {
+    fun tilesInSector(sector: Sector): Long {
         val deltaLat = this.sector.deltaLatitude / matrixHeight
         val deltaLon = this.sector.deltaLongitude / matrixWidth
         val firstRow = Tile.computeRow(deltaLat, sector.minLatitude, this.sector.minLatitude)
         val lastRow = Tile.computeLastRow(deltaLat, sector.maxLatitude, this.sector.minLatitude)
         val firstCol = Tile.computeColumn(deltaLon, sector.minLongitude, this.sector.minLongitude)
         val lastCol = Tile.computeLastColumn(deltaLon, sector.maxLongitude, this.sector.minLongitude)
-        return (lastRow - firstRow + 1) * (lastCol - firstCol + 1)
+        return (lastRow - firstRow + 1).toLong() * (lastCol - firstCol + 1).toLong()
     }
 }
