+++ worldwind/src/jvmMain/kotlin/earth/worldwind/layer/TiledImageLayer.kt
@@ -41,7 +41,7 @@ actual open class TiledImageLayer actual constructor(
     @Throws(IllegalStateException::class, IllegalArgumentException::class)
     fun makeLocal(
         sector: Sector, resolution: ClosedRange<Angle>, overrideCache: Boolean = false,
-        scope: CoroutineScope = GlobalScope, onProgress: ((Int, Int, Int) -> Unit)? = null
+            scope: CoroutineScope = GlobalScope, onProgress: ((Long, Long, Long) -> Unit)? = null
     ): Job {
         val imageDecoder = ImageDecoder()
         return tiledSurfaceImage?.launchBulkRetrieval(scope, sector, resolution, onProgress) { imageSource, cacheSource, options ->
